feat: 股票实时统计图 close #71
