+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ChangeLog
 
+## [1.5.6]
+
+- feat: 股票实时统计图
+- feat: 基金历史趋势统计图
+
 ## [1.5.5]
 
 - feat: 新增基金持仓矩形树图
