+++ demo/stocktrend.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Document</title>
+  </head>
+  <body>
+    <iframe
+      src="http://quote.eastmoney.com/basic/full.html?mcid=1.000001"
+      frameborder="0"
+      style="width: 100%; height: 800px"
+    ></iframe>
+  </body>
+</html>
